+++ client/views/nav/notification_item.js
@@ -1,4 +1,4 @@
-Template[getTemplate('notification_item')].helpers({
+Template[getTemplate('notificationItem')].helpers({
   nice_time: function(){
     return moment(this.timestamp).fromNow();
   },
@@ -10,7 +10,7 @@ Template[getTemplate('notification_item')].helpers({
   }
 });
 
-Template[getTemplate('notification_item')].events({
+Template[getTemplate('notificationItem')].events({
   'click .action-link': function(event, instance){
     var notificationId=instance.data._id;
     Notifications.update(