+++ History.md
@@ -3,6 +3,7 @@
 * Added new invite features (thanks [@callmephilip](https://github.com/callmephilip)!)
 * Changed `navItems` to `primaryNav` and added `secondaryNav`. 
 * Added new `themeSettings` object for storing theme-level settings.
+* Notifications is now a nav menu item. 
 
 ## v0.9.2.5 “AccountScope”